Fix for DIRSERVER-1475
git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@919009 13f79535-47bb-0310-9956-ffa450edef68

+++ dsml-parser/src/main/java/org/apache/directory/shared/dsmlv2/request/Dsmlv2Grammar.java
@@ -1938,7 +1938,7 @@ public class Dsmlv2Grammar extends AbstractGrammar implements IGrammar
             {
                 try
                 {
-                    searchRequest.setSizeLimit( Integer.parseInt( attributeValue ) );
+                    searchRequest.setSizeLimit( Long.parseLong( attributeValue ) );
                 }
                 catch ( NumberFormatException e )
                 {
@@ -1947,7 +1947,7 @@ public class Dsmlv2Grammar extends AbstractGrammar implements IGrammar
             }
             else
             {
-                searchRequest.setSizeLimit( 0 );
+                searchRequest.setSizeLimit( 0L );
             }
             // timeLimit
             attributeValue = xpp.getAttributeValue( "", "timeLimit" );
